[ovsp4rt] Enclose match-key assignments in blocks
Ensured that each code sequence that defines a match key is
enclosed in a local block. This limits the scope of the pointer
to the entry in the protobuf and contributes to the readability
of the code.

Added a brief one-line comment to each block, specifying the key
that is being defined.

Skipped running clang-format on this commit, to make it easier
to diff. Note that you need to specify -b to ignore the added
indentation. The file will be formatted in a later commit.

Also added a "called-by:" comment to each ConfigTableEntry
function, documenting the DoConfigTableEntry function that
calls it.

Signed-off-by: Derek Foster <justffoulkes@gmail.com>