[ovsp4rt] Moved DoConfigEntry tests to 'extras' folder (#34)
This CL quarantines the unit tests for the ES2K DoConfig functions.
It does so by moving the tests to `tests/es2k/extras` and placing them
under control of the `EXTRA_TESTS` cmake option, which is disabled by
default. 

The tests increase the coverage statistics, but they provide little 
assurance that the code being executed is correct. There aren't enough 
observable outputs to test (not enough ASSERTs). Running code without 
verifying the outputs is not testing. 

The tests are also coupled to the internal implementations of the
DoConfig functions, which makes them brittle. 

These tests do provide some useful information, but they are of low 
quality. They should only be run by someone who understands their 
limitations and is prepared to fix them when they break. 

Signed-off-by: Derek Foster <justffoulkes@gmail.com>